+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# 0.5.0
+
+ - added `inspect_name_override` constructor param, useful for dynamic declarations:
+
+    ```ruby
+    IOStruct.new("NN").new.inspect                                 # "<#<Class:0x000000011c45fa20> f0=nil f4=nil>"
+    IOStruct.new("NN", inspect_name_override: "Point").new.inspect # "<Point f0=nil f4=nil>"
+    ```
+
 # 0.4.0
 
  - added `size` class method that returns SIZE constant